requirements.txt
streamlit
openai
pandas
numpy
scikit-learn
python-dotenv
